% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkAnalysisWrapper.R
\name{NetworkAnalysis}
\alias{NetworkAnalysis}
\title{Network Analysis}
\usage{
NetworkAnalysis(
  data = NULL,
  version = "0.97.1",
  betweenness = TRUE,
  bootstrap = FALSE,
  bootstrapParallel = FALSE,
  bootstrapSamples = 0,
  bootstrapType = "nonparametric",
  centralityNormalization = "normalized",
  centralityPlot = FALSE,
  centralityTable = FALSE,
  closeness = TRUE,
  clusteringPlot = FALSE,
  clusteringTable = FALSE,
  colorGroupVariables = list(optionKey = "variable", types = list(), value = list()),
  computedLayoutX = "",
  computedLayoutY = "",
  correlationMethod = "auto",
  criterion = "ebic",
  cut = 0,
  details = FALSE,
  edgeLabelPosition = 0.5,
  edgeLabelSize = 1,
  edgeLabels = FALSE,
  edgePalette = "colorblind",
  edgeSize = 1,
  estimator = "ebicGlasso",
  expectedInfluence = TRUE,
  groupingVariable = list(types = list(), value = ""),
  isingEstimator = "pseudoLikelihood",
  labelAbbreviation = FALSE,
  labelAbbreviationLength = 4,
  labelScale = TRUE,
  labelSize = 1,
  layout = "spring",
  layoutNotUpdated = FALSE,
  layoutSavedToData = FALSE,
  layoutSpringRepulsion = 1,
  layoutX = list(types = list(), value = ""),
  layoutY = list(types = list(), value = ""),
  legend = "allPlots",
  legendSpecificPlotNumber = 1,
  legendToPlotRatio = 0.4,
  manualColor = FALSE,
  manualColorGroups = list(list(name = "Group 1"), list(name = "Group 2")),
  maxEdgeStrength = 0,
  mgmCategoricalVariables = list(types = list(), value = list()),
  mgmContinuousVariables = list(types = list(), value = list()),
  mgmCountVariables = list(types = list(), value = list()),
  mgmVariableTypeShown = "nodeShape",
  minEdgeStrength = 0,
  missingValues = "pairwise",
  nFolds = 3,
  networkPlot = FALSE,
  nodePalette = "colorblind",
  nodeSize = 1,
  plotHeight = 320,
  plotWidth = 480,
  rule = "and",
  sampleSize = "maximum",
  signedNetwork = TRUE,
  split = "median",
  statisticsCentrality = TRUE,
  statisticsEdges = TRUE,
  strength = TRUE,
  thresholdBox = "value",
  thresholdMethod = "sig",
  thresholdValue = 0,
  tuningParameter = 0.5,
  variableNamesShown = "inNodes",
  variables = list(types = list(), value = list()),
  weightedNetwork = TRUE
)
}
\description{
Network Analysis
}
